Move more schemas
