+++ src/components/category.astro
@@ -6,8 +6,7 @@ import TranslatedBlock from "./translated-block.astro";
 import Guest from "./guest.astro";
 import LoadingIndicator from "./loading-indicator.astro";
 import CategoryNavByYear from "./category-nav-by-year.astro";
-import type { Category } from "~/content.config";
-import type { RawProject } from "~/schema";
+import type { Category, RawProject } from "~/schema";
 
 export interface Props {
   language: string;
