+++ src/schema.ts
@@ -61,3 +61,17 @@ export const allYearsSchema = z.array(
     languages: z.array(languageCodeSchema),
   }),
 );
+
+export const categorySchema = z.object({
+  key: z.string(),
+  limit: z.number().optional(),
+  count: z.number().optional(),
+  tags: z.array(z.string()).optional(),
+  excluded: z.array(z.string()).optional(),
+  excludedTags: z.array(z.string()).optional(),
+  availableComments: z.array(z.string()).optional(),
+  guest: z.string().optional(),
+  disabled: z.boolean().optional(),
+});
+
+export type Category = z.infer<typeof categorySchema>;
